Refactor Permiso model to add to_datagrid_dict method for JSON formatting; move get_datagrid_options function to utils for better organization
